[OPP-1530] Refaktorere pub/sub klasser
