+++ frontend-app/src/main/kotlin/no/nav/modialogin/persistence/PersistencePubSub.kt
@@ -1,11 +1,44 @@
 package no.nav.modialogin.persistence
 
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.consumeAsFlow
+import kotlinx.coroutines.launch
+import no.nav.modialogin.Logging
 
 abstract class PersistencePubSub(
     open val channelName: String,
+    private val implementationName: String
 ) {
-    abstract fun startSubscribing(): Flow<String>
-    abstract fun stopSubscribing()
+    private var job: Job? = null
+    protected var channel = Channel<String>()
+    protected var running = false
+    fun startSubscribing(): Flow<String> {
+        doStart()
+        return channel.consumeAsFlow()
+    }
+
+    fun stopSubscribing() {
+        channel.close()
+        doStop()
+        channel = Channel()
+    }
+    private fun doStart() {
+        running = true
+        Logging.log.info("Starting $implementationName subscriber on channel '$channelName'")
+        job = GlobalScope.launch {
+            subscribe()
+        }
+    }
+
+    private fun doStop() {
+        Logging.log.info("Stopping the $implementationName subsriber on channel '$channelName")
+        running = false
+        job?.cancel()
+    }
+
+    abstract suspend fun subscribe(retryInterval: Long = 5000)
     abstract suspend fun publishData(data: String)
 }
